Ensure that datasette can be run with systemd
A few problems were preventing us from running datasette as a systemd
service, and ensuring the dataset updates automatically. For one, uv
wasn't available on that box, meaning some just tasks were failing, and
secondly it's installed only for the "deploy" user, meaning it's
unavailable from the systemd script.

This has been fixed by ensuring ansible correctly provisions uv, and
splitting the responsiblility for running the datasette server and
updataing the dataset into two places - a systemd script, and a cronjob,
respectively. The former has no need for uv, and is run by systemd, the
other does require uv, and is therefore run by cron as the deploy user,
within a bash shell. This is timed to run daily at 5am, (leaving a
window of time after the daily snapshot is generated at 1.30am by the
admin-portal app).